# docs/swagger.yaml
definitions:
...
  datastore.Card:
    properties:
      approved:
        type: string
      company_id:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      linkToProd:
        type: string
      name:
        type: string
      price:
        type: string
      tags:
        items:
          type: string
        type: array
      views:
        type: integer
    type: object
  datastore.Company:
    properties:
      cards:
        items:
          $ref: '#/definitions/datastore.Card'
        type: array
      deletedAt:
        type: string
      id:
        type: string
      image:
        type: string
      isDel:
        type: integer
      linkToCompany:
        type: string
      mail:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  utility.Message:
    properties:
      text:
        type: string
    type: object
host: http://localhost:8081
info:
  contact: {}
  description: Nenavijy swagger
  license:
    name: BSD-3
  title: pay-with-crypto API
  version: "1.0"
paths:
...
  /card/create:
    post:
      consumes:
      - application/json
      description: Card create
      parameters:
      - description: Card data
        in: body
        name: card_data
        required: true
        schema:
          properties:
            description:
              type: string
            linkToProd:
              type: string
            name:
              type: string
            price:
              type: string
            tags:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Card successful created
          schema:
            $ref: '#/definitions/datastore.Card'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utility.Message'
        "409":
          description: Already created
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      security:
      - accessToken: []
      tags:
      - Card
  /card/delete:
    delete:
      consumes:
      - application/json
      description: Card delete
      parameters:
      - description: Card data
        in: body
        name: card_data
        required: true
        schema:
          properties:
            id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: Card successful deleted
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utility.Message'
        "403":
          description: Other owner
          schema:
            $ref: '#/definitions/utility.Message'
        "404":
          description: No card
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      security:
      - accessToken: []
      tags:
      - Card
  /card/edit:
    put:
      consumes:
      - application/json
      description: Card edit
      parameters:
      - description: Card data
        in: body
        name: card_data
        required: true
        schema:
          properties:
            description:
              type: string
            linkToProd:
              type: string
            name:
              type: string
            price:
              type: string
            tags:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Card successful edited
          schema:
            $ref: '#/definitions/utility.Message'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utility.Message'
        "403":
          description: Other owner
          schema:
            $ref: '#/definitions/utility.Message'
        "404":
          description: No card
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      security:
      - accessToken: []
      tags:
      - Card
  /card/getLogo:
    get:
      consumes:
      - application/json
      description: Get card logo
      parameters:
      - description: Card id
        in: query
        name: cardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Card logo link
          schema:
            $ref: '#/definitions/utility.Message'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utility.Message'
        "404":
          description: No card
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Card
  /card/search:
    get:
      consumes:
      - application/json
      description: Search card
      parameters:
      - description: Card name
        in: query
        name: name
        type: string
      - description: Card tags
        in: query
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/datastore.Card'
            type: array
        "404":
          description: No cards
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Card
  /card/search/id:
    get:
      consumes:
      - application/json
      description: Search card by id
      parameters:
      - description: Card id
        in: query
        name: cardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datastore.Card'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utility.Message'
        "403":
          description: Card owner was banned
          schema:
            $ref: '#/definitions/utility.Message'
        "404":
          description: No card
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Card
  /card/uploadLogo:
    post:
      consumes:
      - application/json
      description: Card logo uploader
      parameters:
      - description: Card id
        in: query
        name: cardId
        required: true
        type: string
      - description: Logo image
        in: formData
        name: cardLogo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: Card successful uploaded
        "400":
          description: Invalid image
          schema:
            $ref: '#/definitions/utility.Message'
        "403":
          description: Card owner was banned
          schema:
            $ref: '#/definitions/utility.Message'
        "404":
          description: No card
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      security:
      - accessToken: []
      tags:
      - Card
  /company/search/id:
    get:
      consumes:
      - application/json
      description: Search company by id
      parameters:
      - description: Company id
        in: query
        name: companyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datastore.Company'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utility.Message'
        "403":
          description: Card owner was banned
          schema:
            $ref: '#/definitions/utility.Message'
        "404":
          description: No card
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Company
  /company/uploadLogo:
    post:
      consumes:
      - application/json
      description: Company logo uploader
      parameters:
      - description: Logo image
        in: formData
        name: companyLogo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: Company logo successful uploaded
        "400":
          description: Invalid request, provide companyLogo
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      security:
      - accessToken: []
      tags:
      - Company
securityDefinitions:
  string:
    in: header
    name: accessToken
    type: apiKey
swagger: "2.0"